requirements.txt
flask==3.0.3
flask-cors==4.0.1
bcrypt==4.1.3
psycopg2-binary==2.9.9
pyjwt==2.8.0
gunicorn==22.0.0
pillow==10.3.0
python-dotenv==1.0.1